requirements.txt
streamlit
tiktoken
transformers
openai
pandas
numpy
langchain
langchain-community
unstructured
faiss-cpu
sentence-transformers
 